move gru_model from API folder to root
