+++ generate.py
@@ -1,5 +1,4 @@
 import os
-import codecs
 import argparse
 
 import numpy as np
@@ -18,6 +17,12 @@ def softmax(output):
     output /= np.exp(output * theta)
     output /= np.sum(output)
     return output
+
+def highest_epoch(opt):
+    epoch_list = os.listdir(opt.save_dir)
+    epoch_list = [int(epoch[4:-4]) for epoch in epoch_list]
+    highest_epoch = sorted(epoch_list)[-1]
+    return highest_epoch
 
 def word_discriminator(output):
     return output[-1,:,:].squeeze().max(dim=0)[1].item()
@@ -91,6 +96,9 @@ def sample_model(opt, model, hidden=None, prime_text=" ", length=1000, text=True
 if __name__ == "__main__":
     # get the information from argparse
     parser = argparse.ArgumentParser("Add information about making samples")
+   
+    parser.add_argument('--run', type=bool, default=False,
+                        help="if you run main method")
     parser.add_argument('--epoch', type=int, default=None,
                         help="The epoch of saved model your gonna load")
     parser.add_argument('--prime', type=str, default=" ",
@@ -101,87 +109,89 @@ if __name__ == "__main__":
                         help="resume previous hidden state?")
     
     args = parser.parse_args()                         
-    assert args.epoch != None, "The epoch must be entered! --epoch [int]"
     
-    # Directory
-    # - Where is trainable data
-    # - Where to save the model parameter
-    # - Where to save the generated text
-    data_dir = "data/"
-    save_dir = "save/"
-    gen_dir = "generate/"
-
-    # Choose the hyperparameter at here!
-    ratio = 0.9
-    num_layers = 2
-    hidden_size = 2048
-    embedding_size = 2048
-    cuda = True if torch.cuda.is_available() else False
-    batch_size = 1 # Change 1 because this is generating
-    seq_len = 50
-    num_epochs = 100
-    save_every = 50
-    print_every = 10
-    valid_every = 20 # test the valid data when batch step is (int)
-    grad_clip = 5.
-    learning_rate = 0.001
-
-    # Store every options to opt class data structure
-    opt = opt(data_dir=data_dir,
-              save_dir=save_dir,
-              gen_dir=gen_dir,
-              ratio = ratio,
-              num_layers=num_layers,
-              hidden_size=hidden_size,
-              embedding_size=embedding_size,
-              cuda=cuda,
-              batch_size=batch_size,
-              seq_len = seq_len,
-              num_epochs=num_epochs,
-              save_every=save_every,
-              print_every=print_every,
-              valid_every=valid_every,
-              grad_clip=grad_clip,
-              learning_rate=learning_rate)
-    
-    # load the vocab data
-    with open('vocab/vocab.pkl', 'rb') as f:
-        vocab = pickle.load(f)
-
-    # load the hidden data if resume is true
-    if args.resume:
-        with open(os.path.join(opt.gen_dir, "hidden.pkl"), 'rb') as f:
-            h = pickle.load(f)
-    else:
-        h = None
+    if args.run:
+        assert args.epoch != None, "The epoch must be entered! --epoch [int]"
         
-    # Store vocabulary to the option
-    opt.vocab_size = vocab['vocab_size']
-    opt.vocab_itoc = vocab['vocab_itoc']
-    opt.vocab_ctoi = vocab['vocab_ctoi']
+        # Directory
+        # - Where is trainable data
+        # - Where to save the model parameter
+        # - Where to save the generated text
+        data_dir = "data/"
+        save_dir = "save/"
+        gen_dir = "generate/"
     
-    # make model
-    model = gru(opt)
-    if opt.cuda:
-        model = model.cuda()
-
-    # load saved torch data
-    save_path = os.path.join(opt.save_dir, "gru_{0}.pkl".format(args.epoch))
-
-    # load saved model
-    model.load_state_dict(torch.load(save_path))
-
-    result, h = sample_model(opt=opt,
-                          model=model,
-                          hidden=h,
-                          prime_text=args.prime,
-                          length=args.len)
-
-    # store the hidden state
-    with open(os.path.join(opt.gen_dir, "hidden.pkl"), 'wb') as f:
-        pickle.dump(h, f)
-
-    print(result)
+        # Choose the hyperparameter at here!
+        ratio = 0.9
+        num_layers = 2
+        hidden_size = 2048
+        embedding_size = 2048
+        cuda = True if torch.cuda.is_available() else False
+        batch_size = 1 # Change 1 because this is generating
+        seq_len = 50
+        num_epochs = 100
+        save_every = 50
+        print_every = 10
+        valid_every = 20 # test the valid data when batch step is (int)
+        grad_clip = 5.
+        learning_rate = 0.001
+    
+        # Store every options to opt class data structure
+        opt = opt(data_dir=data_dir,
+                  save_dir=save_dir,
+                  gen_dir=gen_dir,
+                  ratio = ratio,
+                  num_layers=num_layers,
+                  hidden_size=hidden_size,
+                  embedding_size=embedding_size,
+                  cuda=cuda,
+                  batch_size=batch_size,
+                  seq_len = seq_len,
+                  num_epochs=num_epochs,
+                  save_every=save_every,
+                  print_every=print_every,
+                  valid_every=valid_every,
+                  grad_clip=grad_clip,
+                  learning_rate=learning_rate)
+        
+        # load the vocab data
+        with open('vocab/vocab.pkl', 'rb') as f:
+            vocab = pickle.load(f)
+    
+        # load the hidden data if resume is true
+        if args.resume:
+            with open(os.path.join(opt.gen_dir, "hidden.pkl"), 'rb') as f:
+                h = pickle.load(f)
+        else:
+            h = None
+            
+        # Store vocabulary to the option
+        opt.vocab_size = vocab['vocab_size']
+        opt.vocab_itoc = vocab['vocab_itoc']
+        opt.vocab_ctoi = vocab['vocab_ctoi']
+        
+        # make model
+        model = gru(opt)
+        if opt.cuda:
+            model = model.cuda()
+    
+        # load saved torch data
+        save_path = os.path.join(opt.save_dir, "gru_{0}.pkl".format(args.epoch))
+    
+        # load saved model
+        model.load_state_dict(torch.load(save_path))
+    
+        result, h = sample_model(opt=opt,
+                              model=model,
+                              hidden=h,
+                              prime_text=args.prime,
+                              length=args.len)
+    
+        # store the hidden state
+        with open(os.path.join(opt.gen_dir, "hidden.pkl"), 'wb') as f:
+            pickle.dump(h, f)
+    
+        print(result)
