+++ multi_train.py
@@ -1,6 +1,5 @@
 import os
 import pickle
-import numpy as np
 import torch
 from tqdm import tqdm
